pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "largeliterarymodels"
version = "0.3.0"
description = "LLMs for digital humanities: structured data extraction from unstructured texts"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dynamic = ["dependencies"]
authors = [
    {name = "Ryan Heuser"},
]
keywords = ["llm", "digital-humanities", "nlp", "structured-data", "pydantic", "anthropic", "openai", "gemini"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
]

[project.urls]
Homepage = "https://github.com/quadrismegistus/largeliterarymodels"
Repository = "https://github.com/quadrismegistus/largeliterarymodels"
Issues = "https://github.com/quadrismegistus/largeliterarymodels/issues"

[project.scripts]
litmod = "largeliterarymodels.cli.main:main"

[project.optional-dependencies]
dev = ["pytest", "pytest-cov"]
lltk = ["lltk-dh"]
annotate = ["fh-pydantic-form>=0.3", "uvicorn"]
analysis = ["scipy>=1.9", "numpy", "pandas"]

[tool.setuptools.dynamic]
dependencies = {file = ["requirements.txt"]}

[tool.setuptools.packages.find]
include = ["largeliterarymodels*"]
